refactor(job): moved some logic from controller to useCase

+++ job/src/useCases/recruiter/delete-job.ts
@@ -1,3 +1,4 @@
+import { ForbiddenError, NotFoundError } from "@abijobportal/common";
 import { IDependenciesData } from "../../frameworks/types/dependencyInterface";
 
 export = (dependencies: IDependenciesData) => {
@@ -7,8 +8,13 @@ export = (dependencies: IDependenciesData) => {
 		throw new Error("jobRepository should exist in dependencies");
 	}
 
-	const execute = (jobId: string) => {
-		return jobRepository.deleteJob(jobId);
+	const execute = async(jobId: string, recruiterId: string) => {
+		const job = await jobRepository.getAJob(jobId);
+		if(!job) throw new NotFoundError();
+		if(recruiterId !== job.recruiterId.toString()){
+			throw new ForbiddenError("You cannot modify others job")
+		}
+		return await jobRepository.deleteJob(jobId);
 	};
 
 	return { execute };
